+++ .env
@@ -1,3 +1,4 @@
 REACT_APP_DEPLOY_ADDRESS = 0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512
-REACT_APP_VERIFICATION = 24
+REACT_APP_VERIFICATION = 10
 REACT_APP_CONSUMER_ADDRESS = 0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0
+REACT_APP_VERIFIER_ADDRESS = 0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9
