+++ src/components/index.js
@@ -3,11 +3,17 @@ export { default as DeleteEntry } from "./userRegistry/DeleteEntry";
 export { default as GetEntry } from "./userRegistry/GetEntry";
 export { default as GetTotalEntries } from "./userRegistry/GetTotalEntries";
 export { default as UpdateEntry } from "./userRegistry/UpdateEntry";
+
 export { default as InsertConsumer } from "./consumerRegistry/InsertConsumer";
 export { default as GetConsumer } from "./consumerRegistry/GetConsumer";
 export { default as UpdateConsumer } from "./consumerRegistry/UpdateConsumer";
 export { default as DeleteConsumer } from "./consumerRegistry/DeleteConsumer";
 export { default as VerifyAttestationWithSig } from "./consumerRegistry/VerifyAttestationWithSig";
 export { default as TotalConsumers } from "./consumerRegistry/TotalConsumers";
+
+export { default as InsertVerifier } from "./verifierRegistry/InsertVerifier";
+export { default as DeleteVerifier } from "./verifierRegistry/DeleteVerifier";
+export { default as GetVerifier } from "./verifierRegistry/GetVerifier";
+
 
 export { default as QRCode } from "./userRegistry/qrcode";
